requirements.txt
rich
pandas
pillow
pyfiglet
openpyxl
db-dtypes
python-dotenv

langgraph
langchain
langchain-core
langchain-ollama
langchain-google-genai
